Fix mixed precision: index strand wrote bf16 into a float32 buffer (0.8.1 / ihelix 0.5.1)
On a GPU the model runs under bf16 autocast. The residual stream stays
float32 while attention products come out in bf16, and IndexAttention
wrote its bf16 result into a float32 buffer with an indexed assignment,
which PyTorch refuses: "Index put requires the source and destination
dtypes match". Every test ran in float32 on a CPU, so the path was never
executed until the first real run.

- The buffer now takes the dtype of the attention result.
- The mixed-precision context is fresh per use; "bf16" is honoured on
  any device so the GPU path can be tested on a CPU; "auto" is bf16 on
  CUDA and full precision elsewhere; a disabled context is a no-op
  rather than autocast(enabled=False), which switched off an outer one.
- Rollout training scales fp16 gradients (GradScaler) on cards without
  bf16.
- Tests run the model (bf16 and fp16), the trainer, the rollout
  ensemble, the forecast and the storm features under autocast; they
  fail with the reported error without the fix. The whole cell was run
  with bf16 forced everywhere.

+++ ihelix/src/ihelix/attention.py
@@ -359,8 +359,12 @@ class IndexAttention(nn.Module):
         attended = attended.permute(0, 3, 4, 1, 2, 5).reshape(
             batch, num_clusters * cluster_size, proj.num_kv_heads, proj.groups, proj.head_dim
         )
+        # The buffer takes the dtype of what is written into it, not of the residual stream: under
+        # autocast the stream stays float32 while the attention product comes out in bf16/fp16, and an
+        # indexed write between the two is an error rather than a cast.
         output = torch.zeros(
-            batch, grid.num_points, proj.num_kv_heads, proj.groups, proj.head_dim, dtype=x.dtype, device=x.device
+            batch, grid.num_points, proj.num_kv_heads, proj.groups, proj.head_dim,
+            dtype=attended.dtype, device=x.device,
         )
         flat_order = grid.order
         keep = grid.cluster_valid.reshape(-1)
